+++ package.json
@@ -32,6 +32,7 @@
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "globals": "^16.3.0",
     "prettier": "^3.6.2",
+    "sass": "^1.89.2",
     "sass-embedded": "^1.89.2",
     "terser": "^5.43.1",
     "vite": "^7.0.3"
