+++ src/components/PostNav/EmojiBar.module.scss
@@ -1,0 +1,13 @@
+.icon {
+  color: #fff;
+  background-color: rgba($color: #000000, $alpha: 0.54);
+  border-radius: 32px;
+  padding: 8px 14px;
+}
+
+.arrow {
+  transition: transform 0.3s ease;
+}
+.arrow.rotate {
+  transform: rotate(180deg);
+}
