+++ src/components/PostNav/EmojiShare.module.scss
@@ -1,0 +1,5 @@
+.buttonBorder {
+  border: 1px solid $gray300;
+  border-radius: 6px;
+  padding: 17px 8px;
+}
