+++ src/components/PostNav/ShareBar.jsx
@@ -1,8 +1,10 @@
+import shareIcon from '@/assets/images/shareIcon.png';
+
 function ShareBar() {
   return (
     <div>
       <button>
-        <img src='' alt='공유 아이콘' />
+        <img src={shareIcon} alt='공유 아이콘' />
       </button>
     </div>
   );
