--- src/components/paperHeader/EmojiBar.jsx
@@ -1,14 +1,0 @@
-function EmojiBar({ iconCount = 0 }) {
-  return (
-    <div className='Icon'>
-      <ol>
-        <li>{iconCount}</li>
-        <li>{iconCount}</li>
-        <li>{iconCount}</li>
-      </ol>
-      <img src='' alt='allowDown' />
-    </div>
-  );
-}
-
-export default EmojiBar;
