--- src/components/paperHeader/EmojiDropdown.jsx
@@ -1,12 +1,0 @@
-function EmojiDropdown() {
-  return (
-    <div>
-      <button>
-        <img src='' alt='이모지추가아이콘' />
-        추가
-      </button>
-    </div>
-  );
-}
-
-export default EmojiDropdown;
