+++ src/layouts/PostPageLayout.jsx
@@ -3,15 +3,14 @@ import { Outlet } from 'react-router-dom';
 import Header from '@/components/layout/Header';
 import useIsMobile from '@/hooks/useIsMobile';
 import styles from '@/layouts/Layout.module.scss';
+import PostNav from '@/pages/PostPage/PostNav';
 
 export default function PostPageLayout() {
   const isMobile = useIsMobile();
   return (
     <>
       {isMobile ? null : <Header />}
-      {
-        // 두번째 Header
-      }
+      <PostNav />
       <div className={styles.container}>
         <Outlet />
       </div>
