--- src/pages/PaperMessagesPage/PaperCardList.jsx
@@ -1,4 +1,0 @@
-function PaperCardList() {
-  return <>작성중</>;
-}
-export default PaperCardList;
