--- src/pages/PaperMessagesPage/PaperMessagesPage.jsx
@@ -1,13 +1,0 @@
-import PaperCardList from '@/pages/PaperMessagesPage/PaperCardList';
-import PaperHeader from '@/pages/PaperMessagesPage/PaperHeader';
-
-function PaperMessagesPage() {
-  return (
-    <main>
-      <PaperHeader />
-      <PaperCardList />
-    </main>
-  );
-}
-
-export default PaperMessagesPage;
