+++ src/pages/PostPage/PostNav.module.scss
@@ -1,0 +1,9 @@
+.recipient {
+  font-size: $font-3xl;
+  font-weight: 700;
+  color: $gray800;
+}
+
+.headerRight {
+  float: right;
+}
